requirements.txt
# Core
python-dotenv==1.0.1

# UI
streamlit==1.32.2

# LLM (Hugging Face – deployment fallback)
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99

# Utilities
requests==2.31.0
tqdm==4.66.1

